Changed first step in retailer setup to make it simpler. Unified SQL scripts.
The first step was changed so that we can use a google spreadsheet to manage the configuration of campaigns, utm parameters and SKUs. This means the retailer does not need to change the scripts each time they want to add a new campaign/skus. The separate SQL scripts were moved to the separate_queries/ folder. The code for each of the step was moved into the retailer_setup.sql file and updated so that it uses the table linked to the spreadsheet.

Change-Id: I6efa263c89c184c4618f01ef9b641f4f2c676941